+++ lib/inc/path_manager.h
@@ -5,10 +5,10 @@
 #define M_PI_2  1.57079632679489661923
 
 typedef struct WayPoint_ {
-    double x;
-    double y;
+    float x;
+    float y;
 
-    WayPoint_(double x, double y)
+    WayPoint_(float x, float y)
     {
         this->x = x;
         this->y = y;
@@ -17,28 +17,28 @@ typedef struct WayPoint_ {
 } WayPoint;
 
 typedef struct Point_ {
-    double x;
-    double y;
-    double yaw;
-    double k;
-    double speed;
+    float x;
+    float y;
+    float yaw;
+    float k;
+    float speed;
 } Point;
 
 
 typedef struct ControlState_
 {
-    double x;
-    double y;
-    double yaw;
-    double steer;
-    double v;
+    float x;
+    float y;
+    float yaw;
+    float steer;
+    float v;
 
     ControlState_()
     {
 
     }
 
-    ControlState_(double x, double y, double yaw, double steer, double v)
+    ControlState_(float x, float y, float yaw, float steer, float v)
     {
         this->x = x;
         this->y = y;
