+++ lib/src/lqr_pid_control.cpp
@@ -1,4 +1,4 @@
-#include "pid_control.h"
+#include "lqr_pid_control.h"
 
 pid_controller::pid_controller()
 {
@@ -10,7 +10,7 @@ pid_controller::pid_controller()
     this->parameter.output = 0;
 }
 
-void pid_controller::init(double p_gain, double i_gain, double d_gain)
+void pid_controller::init(float p_gain, float i_gain, float d_gain)
 {
     this->parameter.p_gain = p_gain;
     this->parameter.i_gain = i_gain;
@@ -20,12 +20,12 @@ void pid_controller::init(double p_gain, double i_gain, double d_gain)
     this->parameter.output = 0;
 }
 
-void pid_controller::set_target(double target)
+void pid_controller::set_target(float target)
 {
     this->parameter.target = target;
 }
 
-double pid_controller::calculate(double current_value)
+float pid_controller::calculate(float current_value)
 {
     this->parameter.error = this->parameter.target - current_value;
