+++ lib/src/pid_steer_control.cpp
@@ -8,11 +8,11 @@
 
 #define STANLEY_TAG "STANLEY_RUNNER_TAG"
 
-const double MAX_STEER = 45.0 * M_PI / 180.0;
-const double MAX_SPEED = 10.0 / 3.6;
-const double WB = 0.41;                      //앞 뒤 바퀴 사이 거리
+const float MAX_STEER = 45.0 * M_PI / 180.0;
+const float MAX_SPEED = 10.0 / 3.6;
+const float WB = 0.41;                      //앞 뒤 바퀴 사이 거리
 
-double k = 0.5; // gain
+float k = 0.5; // gain
 
 pid_steer_control::pid_steer_control()
 {
@@ -25,7 +25,7 @@ pid_steer_control::~pid_steer_control()
 
 }
 
-void pid_steer_control::generate_spline(ControlState init_state, std::vector<WayPoint> waypoints, double target_speed, double ds)
+void pid_steer_control::generate_spline(ControlState init_state, std::vector<WayPoint> waypoints, float target_speed, float ds)
 {
     CubicSpline2D spline(waypoints);
     this->points = spline.generate_spline_course(target_speed, ds); // get spline points
@@ -71,12 +71,12 @@ void pid_steer_control::add_course(ControlState init_state, std::vector<Point> p
 int pid_steer_control::calculate_nearest_index(ControlState state, std::vector<Point> points, int pind)
 {
     const int N_IND_SEARCH = 5;
-    double min = 10000;
+    float min = 10000;
     uint32_t min_point_index = 0;
     for (uint32_t i = pind; i < (pind + N_IND_SEARCH); i++) {
-        double dx = state.x - points[i].x;
-        double dy = state.y - points[i].y;
-        double point_to_distance = dx * dx + dy * dy;
+        float dx = state.x - points[i].x;
+        float dy = state.y - points[i].y;
+        float point_to_distance = dx * dx + dy * dy;
         if (min > point_to_distance) {
             min = point_to_distance;
             min_point_index = i;
@@ -88,21 +88,21 @@ int pid_steer_control::calculate_nearest_index(ControlState state, std::vector<P
     return min_point_index;
 }
 
-int pid_steer_control::calculate_target_index(ControlState state, std::vector<Point> points, int pind, double& err_front_axel)
+int pid_steer_control::calculate_target_index(ControlState state, std::vector<Point> points, int pind, float& err_front_axel)
 {
     const int N_IND_SEARCH = 5;
-    double min = 10000;
-    double target_dx = 0;
-    double target_dy = 0;
+    float min = 10000;
+    float target_dx = 0;
+    float target_dy = 0;
     uint32_t min_point_index = 0;
 
-    double fx = state.x + WB * std::cos(state.yaw);
-    double fy = state.y + WB * std::sin(state.yaw);
+    float fx = state.x + WB * std::cos(state.yaw);
+    float fy = state.y + WB * std::sin(state.yaw);
 
     for (uint32_t i = pind; i < (pind + N_IND_SEARCH); i++) {
-        double dx = fx - points[i].x;
-        double dy = fy - points[i].y;
-        double point_to_distance = dx * dx + dy * dy;
+        float dx = fx - points[i].x;
+        float dy = fy - points[i].y;
+        float point_to_distance = dx * dx + dy * dy;
         if (min > point_to_distance) {
             min = point_to_distance;
             target_dx = dx;
@@ -113,8 +113,8 @@ int pid_steer_control::calculate_target_index(ControlState state, std::vector<Po
             break;
         }
     }
-    double front_axle_vec_x = -std::cos(state.yaw + M_PI / 2);
-    double front_axle_vec_y = -std::sin(state.yaw + M_PI / 2);
+    float front_axle_vec_x = -std::cos(state.yaw + M_PI / 2);
+    float front_axle_vec_y = -std::sin(state.yaw + M_PI / 2);
 
     err_front_axel = target_dx * front_axle_vec_x + target_dy * front_axle_vec_y;
 
@@ -124,7 +124,7 @@ int pid_steer_control::calculate_target_index(ControlState state, std::vector<Po
 void pid_steer_control::smooth_yaw(std::vector<Point> &points)
 {
     for (uint32_t i = 0; i < points.size() - 1; i++) {
-        double diff_yaw = points[i + 1].yaw - points[i].yaw;
+        float diff_yaw = points[i + 1].yaw - points[i].yaw;
 
         while (diff_yaw >= M_PI_2) {
             points[i + 1].yaw -= M_PI * 2.0;
@@ -138,32 +138,32 @@ void pid_steer_control::smooth_yaw(std::vector<Point> &points)
     }
 }
 
-int pid_steer_control::pid_steering_control(ControlState state, double& steer)
+int pid_steer_control::pid_steering_control(ControlState state, float& steer)
 {
-    double e = 0;
+    float e = 0;
     int current_target_ind = this->calculate_target_index(state, this->points, this->target_ind, e);
 
     if (current_target_ind < this->target_ind)
         current_target_ind = this->target_ind;
 
-    double th_e = pi_2_pi(this->points[current_target_ind].yaw - state.yaw);
+    float th_e = pi_2_pi(this->points[current_target_ind].yaw - state.yaw);
 
-    double L = WB;
-    double Lf = 1;
-    
+    float L = WB;
+    float Lf = 1;
 
-    double steer_delta = std::atan2(k * e, state.v);
+
+    float steer_delta = std::atan2(k * e, state.v);
 
     steer = th_e + steer_delta;
 
     return current_target_ind;
 }
 
-bool pid_steer_control::update(double dt) {
+bool pid_steer_control::update(float dt) {
     // dt 저장
-    double calculated_steer = 0;
-    static double pe = 0;
-    static double pth_e = 0;
+    float calculated_steer = 0;
+    static float pe = 0;
+    static float pth_e = 0;
     this->dt = dt;
 
     if (dt == 0) {
@@ -173,11 +173,11 @@ bool pid_steer_control::update(double dt) {
     this->target_ind = pid_steering_control(this->state, calculated_steer);
     // PID로 가속도 값 계산
     this->path_accel_pid.set_target(this->points[this->target_ind].speed);
-    double calculated_accel = this->path_accel_pid.calculate(this->state.v);
+    float calculated_accel = this->path_accel_pid.calculate(this->state.v);
 
     // state update
     this->state = this->update_state(this->state, calculated_accel, calculated_steer, this->dt);
-    double state_to_goal_distance = sqrt(pow(this->goal_state.x - this->state.x, 2) + pow(this->goal_state.y - this->state.y, 2));
+    float state_to_goal_distance = sqrt(pow(this->goal_state.x - this->state.x, 2) + pow(this->goal_state.y - this->state.y, 2));
 
     if (state_to_goal_distance < 1 && (this->target_ind > this->points.size()/2)) {
         // finish
@@ -186,21 +186,21 @@ bool pid_steer_control::update(double dt) {
     return false;
 }
 
-double pid_steer_control::calculate_error()
+float pid_steer_control::calculate_error()
 {
-    double e = 0;
-    double distance1 = 100;
-    double distance2 = 100;
+    float e = 0;
+    float distance1 = 100;
+    float distance2 = 100;
     int temp_target_index = this->target_ind - 10;
     if (temp_target_index < 0) {
         temp_target_index = 0;
     }
     int nearest_index = this->calculate_nearest_index(state, this->points, temp_target_index);
 
-    auto cal_diatance_between_line_to_point = [](double x1, double y1, double x2, double y2, double stand_x, double stand_y ) -> double{
-        double a = (y1 - y2) / (x1 - x2);
-        double b = -1;
-        double c = y1 - a * x1;
+    auto cal_diatance_between_line_to_point = [](float x1, float y1, float x2, float y2, float stand_x, float stand_y ) -> float{
+        float a = (y1 - y2) / (x1 - x2);
+        float b = -1;
+        float c = y1 - a * x1;
 
         return abs(a * stand_x + b * stand_y + c) / sqrt(a*a + 1);
     };
@@ -218,7 +218,7 @@ double pid_steer_control::calculate_error()
     return distance1 < distance2 ? distance1 : distance2;
 }
 
-ControlState pid_steer_control::update_state(ControlState state, double accel, double steer_delta, double dt)
+ControlState pid_steer_control::update_state(ControlState state, float accel, float steer_delta, float dt)
 {
     if (steer_delta > MAX_STEER) {
         steer_delta = MAX_STEER;
